Added validation for Java reserved keywords, blank lines, strings not matching java method pattern
